complex and reald modes in one demo file
